pyproject.toml
[project]
name = "pms-demo-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "fish-chuan",email = "newdavidlin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (==5.2)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
